create an init function.
it refuses to init if there's no GWLOG_PATH in environment variable.

it only called once on socket creation, this imply it will never accept-

or process a stream of data that it didn't control, only watch the one that-

tracked from socket call, the beginning of TCP connection sequence.

this changes inspired by sir alviro's source code.